changed some navigations
