+++ src/firebase/UserFirebase.ts
@@ -1,56 +1,74 @@
-// Import necessary functions from Firebase modules
-import { collection, doc, setDoc } from "firebase/firestore";
+// Import the functions you need from the SDKs you need
+import { doc, setDoc, getDoc, updateDoc } from "firebase/firestore";
 import {
     signInWithEmailAndPassword,
     signOut,
     createUserWithEmailAndPassword
 } from "firebase/auth";
-
-// Import auth and db instances from the Firebase configuration file
 import { auth, db } from "./firebaseConfig";
 
-// Define a TypeScript interface for the user object
 type UserObj = {
+    id: string;
     name: string;
-    email: string;
+    email?: string;
     number: string;
     address: string;
 };
 
-// Function to log in the user using Firebase Auth
 export const login = async (email: string, password: string) => {
     await signInWithEmailAndPassword(auth, email, password);
 };
 
-// Function to log out the user using Firebase Auth
 export const signout = async () => {
     await signOut(auth);
 };
 
-// Function to register a new user using Firebase Auth
 export const register = async (email: string, password: string) => {
-    await createUserWithEmailAndPassword(auth, email, password);
+    return createUserWithEmailAndPassword(auth, email, password);
 };
 
-// Function to add a new user to the Firestore database
-export const addUser = async ({ name, email, number, address }: UserObj) => {
+export const addUser = async ({
+    id,
+    name,
+    email,
+    number,
+    address
+}: UserObj) => {
     try {
-        // Get a reference to the "users" collection in Firestore
-        const reference = collection(db, "users");
-
-        // Create a new document in the "users" collection with a unique ID
-        const newDocRef = doc(reference);
-
-        // Set the document data to the provided user information
-        await setDoc(newDocRef, {
-            id: newDocRef.id,
+        await setDoc(doc(db, "users", id), {
+            id,
             name,
             email,
             number,
             address
         });
     } catch (error) {
-        // Log any errors that occur while adding the user to the database
+        console.log(error.message);
+    }
+};
+
+export const fetchOne = async (id: string) => {
+    try {
+        const userDoc = doc(db, "users", id);
+        const user = (await getDoc(userDoc)).data();
+        if (!user) {
+            console.log("No such document");
+        } else {
+            return user;
+        }
+    } catch (error) {
+        console.log(error.message);
+    }
+};
+
+export const updateUserDoc = async ({ id, name, number, address }: UserObj) => {
+    try {
+        await updateDoc(doc(db, "users", id), {
+            name,
+            number,
+            address
+        });
+    } catch (error) {
         console.log(error.message);
     }
 };
