+++ src/firebase/firebaseConfig.ts
@@ -1,9 +1,8 @@
-// Import necessary functions from the Firebase modules
+// Import the functions you need from the SDKs you need
 import { initializeApp } from "firebase/app";
 import { getFirestore } from "firebase/firestore";
 import { getAuth } from "firebase/auth";
 
-// Firebase configuration object containing API keys and other settings
 const firebaseConfig = {
     apiKey: import.meta.env.VITE_API_KEY,
     authDomain: import.meta.env.VITE_AUTH_DOMAIN,
@@ -14,11 +13,8 @@ const firebaseConfig = {
     measurementId: import.meta.env.VITE_MEASUREMENT_ID
 };
 
-// Initialize the Firebase app with the provided configuration
 const app = initializeApp(firebaseConfig);
 
-// Get the Firestore instance using the initialized app
 export const db = getFirestore(app);
 
-// Get the Auth instance using the initialized app
 export const auth = getAuth(app);
